requirements.txt
flask 
langchain 
langchain-community 
langchain_openai
faiss-cpu 
flask-cors
flask-restful
python-dotenv